iGPU Spoofed to HD 6000 (#17)
iGPU spoofed to HD 6000 by default in config-base.plist. It was previously thought this was only required for some i3 models but it also applies to some i5 models. It doesn't seem to have any negative effects on devices that do not require it so it's safe to add as a default.

Mentioned new Intel wifi kext in readme. It won't be covered here but is definitely worth a mention.